+++ AnrtdApi/Anrtd.Domain.UnitTests/Entities/ToDoEntityTests.cs
@@ -1,8 +1,6 @@
-﻿using Anrtd.Domain.Entities;
-using Anrtd.Domain.UnitTests.Factories;
-using Moq;
+﻿using Anrtd.Domain.UnitTests.Factories;
+using Anrtd.Domain.UnitTests.Moqs;
 using NUnit.Framework;
-using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
@@ -21,14 +19,17 @@ namespace Anrtd.Domain.UnitTests.Entities
                     "tag1", "tag2", "tag3", "tag4", "tag5"
                 });
 
+                var getTagEntitiesMock = new GetToDoEntitiesMoq();
+
                 await toDo.UpdateTags(
                     new List<string>() { "tag2", "tag4" },
-                    GetTagEntities);
+                    getTagEntitiesMock.Object);
 
                 var actual = toDo.Tags.Select(tag => tag.Id).ToList();
                 var expected = new List<string>() { "tag2", "tag4" };
 
                 CollectionAssert.AreEqual(expected, actual);
+                getTagEntitiesMock.VerifyNoOtherCalls();
             }
 
             [Test]
@@ -39,10 +40,7 @@ namespace Anrtd.Domain.UnitTests.Entities
                     "tag1", "tag2"
                 });
 
-                var getTagEntitiesMock = new Mock<Func<List<string>, Task<List<ToDoTagEntity>>>>();
-                getTagEntitiesMock
-                    .Setup(mock => mock(It.IsAny<List<string>>()))
-                    .Returns((List<string> tagIds) => GetTagEntities(tagIds));
+                var getTagEntitiesMock = new GetToDoEntitiesMoq();
 
                 await toDo.UpdateTags(
                     new List<string>() { "tag1", "tag2", "tag3", "tag4" },
@@ -53,21 +51,7 @@ namespace Anrtd.Domain.UnitTests.Entities
                 var expected = new List<string>() { "tag1", "tag2", "tag3", "tag4" };
 
                 CollectionAssert.AreEqual(expected, actual);
-
-                getTagEntitiesMock.Verify(getTagEntities =>
-                    getTagEntities(It.Is<List<string>>(l => l.Contains("tag3"))),
-                    Times.Once());
-
-                getTagEntitiesMock.Verify(getTagEntities =>
-                   getTagEntities(It.Is<List<string>>(l => l.Contains("tag4"))),
-                   Times.Once());
-            }
-
-            private static Task<List<ToDoTagEntity>> GetTagEntities(List<string> tagIds)
-            {
-                return Task.FromResult(tagIds
-                    .Select(tagId => new ToDoTagEntity() { Id = tagId })
-                    .ToList());
+                getTagEntitiesMock.VerifyWasCalledWithTagIds(new List<string> { "tag3", "tag4" });
             }
         }
     }
